refactor: group agent state modules
